feat(notifications): move notification initialization to RootLayout and restore refetch functions
