+++ src/layout/RootLayout.jsx
@@ -1,7 +1,14 @@
 import { useEffect } from "react";
 import { Outlet, useLocation } from "react-router-dom";
+import { requestPermission, listenToMessages } from "../firebase/service";
 import AOS from "aos";
 import useAuth from "../hooks/useAuth";
+
+import useGetOrder from "../hooks/orders/useGetOrder";
+import useGetProviderOrders from "../hooks/orders/useGetProviderOrders";
+import useGetOrders from "../hooks/orders/useGetOrders";
+import useGetNotifications from "../hooks/settings/useGetNotifications";
+import useGetUserData from "../hooks/user/useGetUserData";
 
 import Header from "../ui/Header";
 import Footer from "../ui/Footer";
@@ -16,6 +23,12 @@ export default function RootLayout() {
   const { isLoading } = useGetHomeSlider();
   const { isLoading: servicesLoading } = useGetServices();
   const { loading } = useAuth();
+
+  const { refetch: refetchOrder } = useGetOrder();
+  const { refetch: refetchOrders } = useGetOrders();
+  const { refetch: refetchUserData } = useGetUserData();
+  const { refetch: refetchNotifications } = useGetNotifications();
+  const { refetch: refetchProviderOrders } = useGetProviderOrders();
 
   useEffect(() => {
     const sections = document.querySelectorAll("section");
@@ -51,6 +64,30 @@ export default function RootLayout() {
       }, 200);
     }
   }, [location]);
+
+  useEffect(() => {
+    const initializeNotifications = async () => {
+      await requestPermission();
+      const unsubscribe = listenToMessages(
+        refetchOrder,
+        refetchOrders,
+        refetchProviderOrders,
+        refetchNotifications,
+        refetchUserData
+      );
+      return () => {
+        if (unsubscribe) unsubscribe();
+      };
+    };
+
+    initializeNotifications();
+  }, [
+    refetchOrder,
+    refetchOrders,
+    refetchProviderOrders,
+    refetchUserData,
+    refetchNotifications,
+  ]);
 
   return loading || isLoading || servicesLoading ? (
     <PageLoader />
